coder-backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "noecoder"
version = "0.1.0"
description = "Personal coding assistant webserver powered by NoeAgent"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Xiaming Chen", email = "chenxm35@gmail.com"}
]
requires-python = ">=3.11"
dependencies = [
    # Core
    "pydantic>=2.0.0",
    "pydantic-settings>=2.12.0",
    "uuid7>=0.1.0",
    # Web framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    # WebSocket
    "python-socketio>=5.11.0",
    "python-engineio>=4.9.0",
    # Git operations
    "gitpython>=3.1.0",
    # Async file I/O
    "aiofiles>=24.1.0",
    # NoeAgent framework
    "noesium>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.27.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.10.0",
]

[project.scripts]
noecoder = "noecoder.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/noecoder"]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
